customer-service-backend/pyproject.toml
[project]
name = "custom-service"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "aiosqlite>=0.22.1",
    "cryptography>=46.0.7",
    "fastapi>=0.116.1",
    "greenlet>=3.4.0",
    "jinja2>=3.1.6",
    "langchain>=1.2.15",
    "langchain-openai>=1.0.1",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
